oncomplate: add model function

+++ app/src/models/perusahaan.js
@@ -1,11 +1,4 @@
-const { connection } = require("../config/database.js");
-
-const perusahaan = {
-  nama_perusahaan: "",
-  no_telp: "",
-  email: "",
-  alamat: "",
-};
+const { requestToDb } = require("./hooks/hooks.js");
 
 const insertPerusahaan = async ({
   nama_perusahaan = "",
@@ -15,27 +8,21 @@ const insertPerusahaan = async ({
 }) => {
   const _sql = `INSERT INTO perusahaan (nama_perusahaan, no_telp, email, alamat) VALUES ('${nama_perusahaan}', '${no_telp}', '${email}', '${alamat}');`;
 
-  try {
-    const [result, fields] = await connection.query(_sql);
-
-    return {
-      affectedRows: result.affectedRows,
-      id: result.insertId,
-    };
-  } catch (error) {
-    return error;
-  }
+  return await requestToDb(_sql);
 };
 
 const getPerusahaan = async () => {
   const _sql = `SELECT * FROM perusahaan;`;
 
-  try {
-    const [result, fields] = await connection.query(_sql);
-    return result;
-  } catch (error) {
-    return error;
-  }
+  return await requestToDb(_sql);
+};
+
+const getPerusahaanById = async (id = 0) => {
+  if (id === 0) return;
+
+  const _sql = `SELECT * FROM perusahaan WHERE id = ${id};`;
+
+  return await requestToDb(_sql);
 };
 
 const updatePerusahaan = async ({
@@ -47,33 +34,23 @@ const updatePerusahaan = async ({
 }) => {
   if (id === 0) return;
 
+  const result = await getPerusahaansById(id);
+  if (result.length === 0) return;
+
   const _sql = `UPDATE perusahaan SET nama_perusahaan = '${nama_perusahaan}', no_telp = '${no_telp}', email = '${email}', alamat = '${alamat}' WHERE id = ${id};`;
 
-  try {
-    const [result, fields] = await connection.query(_sql);
-
-    return {
-      affectedRows: result.affectedRows,
-    };
-  } catch (error) {
-    return error;
-  }
+  return await requestToDb(_sql);
 };
 
 const deletePerusahaan = async (id = 0) => {
   if (id === 0) return;
 
+  const result = await getPerusahaansById(id);
+  if (result.length === 0) return;
+
   const _sql = `DELETE FROM perusahaan WHERE id = ${id};`;
 
-  try {
-    const [result, fields] = await connection.query(_sql);
-
-    return {
-      affectedRows: result.affectedRows,
-    };
-  } catch (error) {
-    return error;
-  }
+  return await requestToDb(_sql);
 };
 
 module.exports = {
@@ -81,4 +58,5 @@ module.exports = {
   getPerusahaan,
   updatePerusahaan,
   deletePerusahaan,
+  getPerusahaanById,
 };
